+++ mods/common/src/index.ts
@@ -59,6 +59,7 @@ export * as Tracer from "./tracer"
 export * as Assertions from "./assertions"
 export * as CommonConnect from "./connect"
 export * as CommonRequester from "./requester"
+export * as CommonResponse from "./response"
 export * as CommonTypes from "./types"
 export * as CommonErrors from "./errors"
 export * as ConnectSchemas from "./connect/schemas"
