+++ mods/common/test/common.unit.test.ts
@@ -22,7 +22,8 @@ import sinonChai from "sinon-chai"
 import { getObjectProto } from "../src/service"
 import { addressCount, getLocalnetIp, isLocalnet } from "../src/ip_utils"
 import { getRedisUrlFromConfig } from "../src/redis"
-import { createUnauthorizedResponse, getCredentials } from "../src/auth"
+import { getCredentials } from "../src/auth"
+import { createUnauthorizedResponse } from "../src/response"
 import { ResponseType } from "../src/types"
 
 const expect = chai.expect
